Cl2r parallel (#15)
* Changed config.py and data_manager.py and created file cl2r_estimater to include parallel mcmc run in pipeline.

* Cleaned up code in cl2r_estimater.py

* Added paths for cl2r step plot to data_manager.py

* Added file for plotting the corner plot of the mcmc for a single sky simulation.

* Added file for comparing the statistics of the estimated value for r + uncertainties for different sky models / combination of cosmological parameters (reproducing Fig. 8 in Woltz et al.).

* Excluded first l bin from the parameter estimation until figuring out why in this first bin Cl_dust < 0.
(+ some minor changes: extracting delta_l directly from the config file instead of computing it and correcting the description in argparse.ArgumentParser())

* Added function plot_all_cornerplots() for plotting the posteriors for all sky simulations on the same corner plot.
Added function plot_spectra_comparison() for plotting the spectra of the observed and estimated power spectra for the different components.

* Modified the function plot_r_statistics() which plots the  r statistics from different posteriors computed from different sky simulations.
The mean and std are now computed for each individual posteriors first. These means and stds of the different posteriors are then averaged to obtain the final statistics on r.

* Changed the CAMB functions used to compute the generic spectra.